requirements.txt
python-vlc
gmusicapi
requests
pygubu